Refactor SessionViewModel
